src/python/requirements.txt
colorlog==6.8.2
dataclass-wizard==0.22.3
Jinja2==3.1.4
MarkupSafe==2.1.5
PyYAML==6.0.1
typing_extensions==4.7.1
pyinstaller